requirements.txt
aiogram==3.3.0
httpx==0.25.0
chromadb==0.4.15
pypdf==4.0.1
python-dotenv==1.0.0
